Refactor because of change of name, and a little of code organization
