+++ member-content-visibility.php
@@ -8,30 +8,25 @@
  * License: GPL2
  * Text Domain: vc-member-container
  */
-namespace VCMembersContainer;
+namespace MemberContentVisibility;
 
 defined('ABSPATH') or die("No script kiddies please!");
 class VCMembersContainer {
 
   public static function init() {
-    //load vc element after js_composer plugin is loaded
-    add_action('plugins_loaded', function(){
-      include ('vc-shortcodes.php');
-      
-      VCShortcodes::init();  
-      
-    });
-    
+    //Add shortcodes
+    include ('shortcodes.php');
+    Shortcodes::init();
+
     add_action( 'admin_init', array(__CLASS__, 'check_visual_composer_plugin' ));
   }
-  
+
   public static function check_visual_composer_plugin(){
     $plugin = "js_composer/js_composer.php";
-    
+
     /*if visual composer plugin is active, then it is enabled the integration*/
     if(is_plugin_active($plugin)){
       include ('vc-integration.php');
-      
       VCIntegration::init();
     }
   }
